test(memory): name simulator suites after the module they cover
Renames each simulator suite to sit beside its subject (diff.test.ts next to
diff.ts) and splits the combined extract/replay guard suite along the module
boundary it already used, so every file under scripts/simulate has a matching
test file. Also drops the changeset: this is repo-local tooling under scripts/,
not a shipped package feature.
