--- .changeset/olive-moons-repeat.md
@@ -1,5 +1,0 @@
----
-'@mastra/memory': patch
----
-
-Add a conversation simulation tool that replays real conversation threads through the Subconscious capture and curation pipeline against a local Postgres, so capture and curation prompt changes can be A/B'd on real data locally instead of deployed and observed over days.
