+++ packages/memory/scripts/simulate/diff.test.ts
