+++ packages/memory/scripts/simulate/drive.test.ts
