+++ packages/memory/scripts/simulate/reconstruct.test.ts
